services/emails/func-emails/requirements.txt
azure-functions
azure-cli
azure-storage-blob
azure-keyvault-secrets
azure-identity
python-dotenv
requests
msal